PoC for a gradle plugin that automates downloading of release artifacts
