+++ .gitignore
@@ -2,7 +2,7 @@ target/
 ### https://raw.github.com/github/gitignore/f2ce448f2ba7a092da05482ceca99209127c0884/Gradle.gitignore
 
 .gradle
-/build/
+build/
 
 # Ignore Gradle GUI config
 gradle-app.setting
