+++ build.gradle
@@ -1,6 +1,7 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This is a general purpose Gradle build.
- * Learn how to create Gradle builds at https://guides.gradle.org/creating-new-gradle-builds/
- */
+import org.apache.commons.releaseverifcation.ReleaseVerificationPlugin
+
+apply plugin: ReleaseVerificationPlugin
+
+distribution {
+    baseUrl = 'https://dist.apache.org/repos/dist/dev/commons/jcs/'
+}
