+++ buildSrc/src/main/groovy/org/apache/commons/releaseverifcation/DistributionExtension.groovy
@@ -1,0 +1,5 @@
+package org.apache.commons.releaseverifcation
+
+class DistributionExtension {
+    String baseUrl
+}
